requirements.txt
# Core dependencies
fastapi>=0.115.0,<1.0.0
pydantic>=2.0.0,<3.0.0
PyJWT>=2.0.0,<3.0.0
SQLAlchemy>=2.0.0,<3.0.0
redis>=5.0.0,<6.0.0
python-dotenv>=1.0.0,<2.0.0
uvicorn>=0.24.0,<1.0.0
httpx>=0.24.0,<1.0.0

# Testing dependencies
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.25.0,<1.0.0
pytest-cov>=4.0.0,<5.0.0
pytest-mock>=3.0.0,<4.0.0
pytest-timeout>=2.0.0,<3.0.0
faker>=19.0.0,<20.0.0

# Development dependencies
black>=24.0.0,<25.0.0
flake8>=7.0.0,<8.0.0
isort>=5.0.0,<6.0.0
mypy>=1.0.0,<2.0.0
pre-commit>=3.0.0,<4.0.0
